refactor: convert constants.js to Typescript
