+++ src/blocks/data.js
@@ -20,7 +20,7 @@
 
 import * as Blockly from "blockly/core";
 import { Categories } from "../categories.js";
-import * as Constants from "../constants.js";
+import * as Constants from "../constants";
 import * as scratchBlocksUtils from "../scratch_blocks_utils.js";
 import { renameVariable } from "../variables";
