+++ src/renderer/renderer.js
@@ -7,7 +7,7 @@
 import * as Blockly from "blockly/core";
 import { Drawer } from "./drawer.js";
 import { RenderInfo } from "./render_info.js";
-import { ConstantProvider } from "./constants.js";
+import { ConstantProvider } from "./constants";
 import { PathObject } from "./path_object.js";
 
 /**
